introduce dbus-ping-systemd and dbus-test-service-systemd
dbus-ping and dbus-test-service are purely libdbus based. The
new client and service are based on the still in development
libsystemd-bus, which is part of the systemd sources. It's a
simple lightweight D-Bus library, which in our preliminary
tests is more than 3 times faster than libdbus when transferring
empty D-Bus messages.

The systemd git repository has been added as a submodule. Don't
forget to 'git submodule update --init'.